[#151]refactor: break,pass 파일 하나로 합침(obj, stmt)
