+++ app/visualize/analysis/stmt/models/flow_control_obj.py
@@ -2,6 +2,13 @@ from dataclasses import dataclass, field
 
 
 @dataclass(frozen=True)
+class BreakStmtObj:
+    id: int
+    expr: str
+    type: str = field(default="flowControl", init=False)
+
+
+@dataclass(frozen=True)
 class PassStmtObj:
     id: int
     expr: str
