--- app/visualize/analysis/stmt/models/flowcontrolobj/break_stmt_obj.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass, field
-
-
-@dataclass(frozen=True)
-class BreakStmtObj:
-    id: int
-    expr: str
-    type: str = field(default="flowControl", init=False)
