--- app/visualize/analysis/stmt/parser/flowcontrol/pass_stmt.py
@@ -1,9 +1,0 @@
-import ast
-
-from app.visualize.analysis.stmt.models.flowcontrolobj.pass_stmt_obj import PassStmtObj
-
-
-class PassStmt:
-    @staticmethod
-    def parse(node: ast.Pass) -> PassStmtObj:
-        return PassStmtObj(id=node.lineno, expr="pass")
