+++ tests/visualize/generator/converter/test_flow_control_converter.py
@@ -1,7 +1,6 @@
 import pytest
 
-from app.visualize.analysis.stmt.models.flowcontrolobj.break_stmt_obj import BreakStmtObj
-from app.visualize.analysis.stmt.models.flowcontrolobj.pass_stmt_obj import PassStmtObj
+from app.visualize.analysis.stmt.models.flow_control_obj import PassStmtObj, BreakStmtObj
 from app.visualize.generator.converter.flow_control_converter import FlowControlConverter
 from app.visualize.generator.models.flow_control_viz import FlowControlViz
