+++ tests/visualize/generator/test_converter_traveler.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 
 import pytest
 
-from app.visualize.analysis.stmt.models.flowcontrolobj.pass_stmt_obj import PassStmtObj
+from app.visualize.analysis.stmt.models.flow_control_obj import PassStmtObj
 from app.visualize.code_visualizer import CodeVisualizer
 from app.visualize.generator.converter.flow_control_converter import FlowControlConverter
 from app.visualize.generator.converter.if_converter import IfConverter
